requirements.txt
fastapi
uvicorn
python-docx
docx2txt
PyMuPDF
llama-cloud-services

